10_Similarity_Search/backend/requirements.txt
fastapi
numpy
python-dotenv
uvicorn
qdrant_client